Feature: Mobile Menu button functionality, Places like button. Refactoring: Menu structure,
The menus for desktop and mobile are refactored. I use a separate file to pull data from it and then run a loop to display the options.
Made a couple more components for reusability. UserAccount and MenuList; Refactored components and improved the quality of codes. Place Card component is also done.