Refactoring: App.css / Style: View All Button / implementation: Newsletter section, Footer section
Rearranged style properties in App.css file. The "View All" button in destinations section was overhauled. It was originally an <a> tag and now I changed it to an actual button tag; that caused the styles to not take effect, therefore changing this element's was necessary.
The Newsletter section is implemented. Also turned the svgs used in this section, into separate components for better abstraction. Footer is implemented.