Refactoring: folder structure and file positions, improve code aesthetics / Fix: issues with lists at footer section in responsive views
There was an issue with lists in the footer of the page that was messing with styles and structure at breakpoints below 1024px; Cause of it was the order of items inside the list, which was resolved by changing it.
I moved the App.css file to public folder for better routing and because it was a global style file. Moved all the images and fonts into public folder and changed all the url paths accordingly; then removed the Images and Fonts folders inside src directory. Changed the order of imports for better Visual DX.